encoding updates

+++ app/models/property.rb
@@ -1,7 +1,9 @@
 # encoding: utf-8
+require 'image'
+
 class Property < ActiveRecord::Base
   
-  attr_accessor :old_district
+  attr_accessor :old_district  
 
   belongs_to :contact
   
@@ -14,17 +16,12 @@ class Property < ActiveRecord::Base
   accepts_nested_attributes_for :contact, reject_if: lambda { |c| c[:name].blank? }, allow_destroy: true
   accepts_nested_attributes_for :images, reject_if: lambda { |c| c[:image].blank? && c[:image_file_name].blank? }, allow_destroy: true
   accepts_nested_attributes_for :property_category_items
-    
+  
   after_save :update_local_values
-  after_commit :update_phones
   
   def item(category)
-    cat = Category.find_by_un(category.to_s)
-    cat = Category.find_by_id(category.to_i) unless cat
-    if cat
-      item = category_items.find_by_category_id(cat.id)
-      item.name if item
-    end
+    item = category_items.find_by_category_id(Category.find_by_un(category.to_s).id)
+    item.name if item
   end
   
   def item_color(category)
@@ -36,9 +33,9 @@ class Property < ActiveRecord::Base
     items = category_items.where(category_id: Category.find_by_un('district').id).pluck(:name)
     items.join(', ') if items.present?
   end
-  
-  def service_types
-    _types = [nil, nil, nil]
+	
+	def service_types	
+		 _types = [nil, nil, nil]
     pairs = category_items.where(category_id: Category.find_by_un('service_type').id).order("property_category_items.weight").pluck("property_category_items.id", "property_category_items.weight", :name)
     pairs.each_with_index do |pair, idx|
       _idx = pair[1].nil? ? idx : pair[1] - 1
@@ -47,6 +44,7 @@ class Property < ActiveRecord::Base
     _types
   end
   
+    
   def self.of_user(user)
     if user.is_main
       all
@@ -71,50 +69,49 @@ class Property < ActiveRecord::Base
     end
     
     Property.where(id: id).update_all(rooms: r, floor: f, floors: fs, address: a, state: s)
-    
-    _images = images
-    
-    if _images.count > 0
-      
-      _district_old = Russian.translit((old_district.present? ? old_district : '-').to_s.force_encoding('utf-8'))
-      _district = Russian.translit((item(:district) || '-').to_s.force_encoding('utf-8'))
-      _landmark_old = Russian.translit(landmark_was.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
-      _landmark = Russian.translit(landmark.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
-      _kee_old = "#{rooms.to_s[/\d+/]}-#{floor.to_s[/\d+/]}-#{floors.to_s[/\d+/]}"
-      _kee = "#{r.to_s[/\d+/]}-#{f.to_s[/\d+/]}-#{fs.to_s[/\d+/]}"
-      _price_old = [price1_was, price2_was, price3_was].reject(&:blank?).join(' ')
-      _price = [price1, price2, price3].reject(&:blank?).join(' ')
-      info_old = "#{_kee_old} #{_landmark_old} #{_price_old}".strip
-      info = "#{_kee} #{_landmark} #{_price}".strip
-      old_path = Rails.root.join('public', 'photos', _district_old, info_old, self.id.to_s)
-      new_path = Rails.root.join('public', 'photos', _district, info, self.id.to_s)
-      
-      _images.each do |image|
-        if image.image?
-          filename = image.short_name  #  image.image_file_name
-          old_path_with_filename = File.join(old_path, filename)
-          
-          if File.exists?(old_path_with_filename) && old_path_with_filename != File.join(new_path, filename)
-            FileUtils.mkdir_p new_path
-            FileUtils.mv old_path_with_filename, new_path
-            
-            # delete old empty folder
-            if Dir[old_path.to_s + '/*'].empty?
-              FileUtils.rm_rf old_path
-              
-              old_parent_path = File.expand_path("..", old_path)
-              if Dir[old_parent_path + "/*"].empty?
-                FileUtils.rm_rf old_parent_path
-              end
-            end
-          end
-          
-          image.save # to update the locally saved image_urls
-        end
-      end
-    end
-  end
-    
+		
+		_images = images
+		
+		if _images.count > 0
+			 _district_old = Russian.translit((old_district.present? ? old_district : '-').to_s.force_encoding('utf-8'))
+			 _district = Russian.translit((item(:district) || '-').to_s.force_encoding('utf-8'))
+			 _landmark_old = Russian.translit(landmark_was.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
+			 _landmark = Russian.translit(landmark.to_s.gsub(/[\x00\/\\:\*\?\"<>\|]/, '').force_encoding('utf-8')).gsub(/[\xC2\xBB|\xC2\xAB]/, '').to_s.strip
+			 _kee_old = "#{rooms.to_s[/\d+/]}-#{floor.to_s[/\d+/]}-#{floors.to_s[/\d+/]}"
+			 _kee = "#{r.to_s[/\d+/]}-#{f.to_s[/\d+/]}-#{fs.to_s[/\d+/]}"
+			 _price_old = [price1_was, price2_was, price3_was].reject(&:blank?).join(' ')
+			 _price = [price1, price2, price3].reject(&:blank?).join(' ')
+			 info_old = "#{_kee_old} #{_landmark_old} #{_price_old}".strip
+			 info = "#{_kee} #{_landmark} #{_price}".strip
+			 old_path = Rails.root.join('public', 'YandexDisk', 'photos', _district_old, info_old, self.id.to_s)
+			 new_path = Rails.root.join('public', 'YandexDisk', 'photos', _district, info, self.id.to_s)
+			
+			_images.each do |image|
+				if image.image?
+					filename = image.short_name  #  image.image_file_name
+					old_path_with_filename = File.join(old_path, filename)
+					
+					if File.exists?(old_path_with_filename) && old_path_with_filename != File.join(new_path, filename)
+						FileUtils.mkdir_p new_path
+						unless File.exists?(File.join(new_path, filename))
+							FileUtils.mv old_path_with_filename, new_path
+							
+							# delete old empty folder
+							if Dir[old_path.to_s + '/*'].empty?
+								FileUtils.rm_rf old_path
+								
+								old_parent_path = File.expand_path("..", old_path)
+								if Dir[old_parent_path + "/*"].empty?
+									FileUtils.rm_rf old_parent_path
+								end
+							end
+						end
+					end
+				end
+			end
+		end
+	end
+  
   def self.update_local_values
     CategoryItem.all.each { |i| i.update_properties }
   end
